www: Add symbol rename config to UI, and use it
